Refactor project to manage products instead of weather forecasts
Deleted WeatherForecast related files as they're no longer needed. A BaseController and ProductsController classes have been introduced to handle product related operations. The ProductService.cs file was revised, and specific service and repository classes related to products were added to the Program.cs. A change in the product repository was also made to depend on the interface rather than the concrete ElasticClient class.