Add Elasticsearch integration and blog feature
This commit introduces the integration of Elasticsearch into the project by including the necessary package and configurations. It also adds the feature to save and retrieve blog data. The blog feature includes a repository for data interactions, service for business logic and appropriate view models for exchange of data between front-end and back-end.