Add search functionality to blog
A new search feature has been implemented in the blog section of the application. This was accomplished by adding a new search method in the BlogRepository and BlogService classes which fetches blog posts based on search text input. Additionally, necessary changes have been made to the BlogController to handle search requests and to the search view to display search results.