Add ECommerceRepository and ECommerce model class
This commit introduces the ECommerceRepository for executing search queries to Elasticsearch. It provides a new method, SearchAsync, which supports dynamic query building based on search filters given in the ECommerceSearchViewModel. Additionally, the ECommerce model class is added; it represents the structure of ecommerce data to be stored and retrieved from Elasticsearch.